+++ cloudformation/batch/aws-batch-jobdefinition_consumableresourcerequirement.go
@@ -18,7 +18,7 @@ type JobDefinition_ConsumableResourceRequirement struct {
 	// Quantity AWS CloudFormation Property
 	// Required: true
 	// See: http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-batch-jobdefinition-consumableresourcerequirement.html#cfn-batch-jobdefinition-consumableresourcerequirement-quantity
-	Quantity int `json:"Quantity"`
+	Quantity int64 `json:"Quantity"`
 
 	// AWSCloudFormationDeletionPolicy represents a CloudFormation DeletionPolicy
 	AWSCloudFormationDeletionPolicy policies.DeletionPolicy `json:"-"`
