+++ cloudformation/greengrassv2/aws-greengrassv2-deployment_systemresourcelimits.go
@@ -18,7 +18,7 @@ type Deployment_SystemResourceLimits struct {
 	// Memory AWS CloudFormation Property
 	// Required: false
 	// See: http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-greengrassv2-deployment-systemresourcelimits.html#cfn-greengrassv2-deployment-systemresourcelimits-memory
-	Memory *int `json:"Memory,omitempty"`
+	Memory *int64 `json:"Memory,omitempty"`
 
 	// AWSCloudFormationDeletionPolicy represents a CloudFormation DeletionPolicy
 	AWSCloudFormationDeletionPolicy policies.DeletionPolicy `json:"-"`
