+++ cloudformation/kafkaconnect/aws-kafkaconnect-customplugin_custompluginfiledescription.go
@@ -18,7 +18,7 @@ type CustomPlugin_CustomPluginFileDescription struct {
 	// FileSize AWS CloudFormation Property
 	// Required: false
 	// See: http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-kafkaconnect-customplugin-custompluginfiledescription.html#cfn-kafkaconnect-customplugin-custompluginfiledescription-filesize
-	FileSize *int `json:"FileSize,omitempty"`
+	FileSize *int64 `json:"FileSize,omitempty"`
 
 	// AWSCloudFormationDeletionPolicy represents a CloudFormation DeletionPolicy
 	AWSCloudFormationDeletionPolicy policies.DeletionPolicy `json:"-"`
